requirements.txt
streamlit
python-decouple
coloredlogs
tqdm
scikit-learn